requirements.txt
asgiref==3.7.2
Django==5.0.1
django-filter==23.5
django-restframework==0.0.1
djangorestframework==3.14.0
Markdown==3.5.2
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
